rona_pkg/rona_pkg/instrument_detector/orientation_estimator/RotNet/requirements.txt
ipython  # interactive notebook
matplotlib>=3.2.2
numpy==1.21.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.10.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
torchsummary
tqdm>=4.64.0
scikit-learn<=1.1.2
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
